moving feed state classes into relevant folder
