+++ Controller/Adminhtml/Datafeed/Runfeed.php
@@ -14,7 +14,7 @@ use Magento\Framework\App\Response\Http;
 use Magento\Framework\App\ResponseInterface;
 use Magento\Framework\Controller\ResultInterface;
 use Magento\Store\Model\StoreManagerInterface;
-use Pureclarity\Core\Model\Feed\Request;
+use Pureclarity\Core\Model\Feed\State\Request;
 
 /**
  * Class Runfeed
