+++ Cron/RunRequestedFeeds.php
@@ -11,7 +11,7 @@ namespace Pureclarity\Core\Cron;
 use Psr\Log\LoggerInterface;
 use Pureclarity\Core\Model\CoreConfig;
 use Pureclarity\Core\Model\Feed\Runner;
-use Pureclarity\Core\Model\Feed\Request;
+use Pureclarity\Core\Model\Feed\State\Request;
 
 /**
  * Class RunScheduled
