+++ Model/Feed/Runner.php
@@ -12,6 +12,10 @@ use Pureclarity\Core\Model\FeedFactory;
 use Pureclarity\Core\Api\StateRepositoryInterface;
 use Psr\Log\LoggerInterface;
 use Pureclarity\Core\Model\CoreConfig;
+use Pureclarity\Core\Model\Feed\State\Running;
+use Pureclarity\Core\Model\Feed\State\RunDate;
+use Pureclarity\Core\Model\Feed\State\Progress;
+use Pureclarity\Core\Model\Feed\State\Error;
 use PureClarity\Api\Feed\Feed;
 use Magento\Framework\Exception\CouldNotDeleteException;
 use Magento\Framework\Exception\CouldNotSaveException;
