+++ Test/Unit/Model/Feed/RunnerTest.php
@@ -14,10 +14,10 @@ use Pureclarity\Core\Model\FeedFactory;
 use Pureclarity\Core\Api\StateRepositoryInterface;
 use Psr\Log\LoggerInterface;
 use Pureclarity\Core\Model\CoreConfig;
-use Pureclarity\Core\Model\Feed\Running;
-use Pureclarity\Core\Model\Feed\RunDate;
-use Pureclarity\Core\Model\Feed\Progress;
-use Pureclarity\Core\Model\Feed\Error;
+use Pureclarity\Core\Model\Feed\State\Running;
+use Pureclarity\Core\Model\Feed\State\RunDate;
+use Pureclarity\Core\Model\Feed\State\Progress;
+use Pureclarity\Core\Model\Feed\State\Error;
 use Pureclarity\Core\Model\Feed\TypeHandler;
 use Pureclarity\Core\Api\FeedDataManagementInterface;
 use Pureclarity\Core\Api\FeedManagementInterface;
